feat: thread dialog ui
