+++ apps/masterbots.ai/app/(browse)/[category]/page.tsx
@@ -1,4 +1,4 @@
-import BrowseList from '@/components/browse/browse-list'
+import BrowseList from '@/components/shared/thread-dialog/thread-list'
 import { BrowseCategoryTabs } from '@/components/browse/browse-category-tabs'
 import { BrowseSearchInput } from '@/components/browse/browse-search-input'
 import { getBrowseThreads, getCategories } from '@/services/hasura'
@@ -13,7 +13,7 @@ export default async function BrowseCategoryPage({
   const categories = await getCategories()
   const categoryId = categories.find(
     c =>
-      c.name.toLowerCase().replace(/\s+/g, '_').replace(/\&/g, 'n') ===
+      c.name.toLowerCase().replace(/\s+/g, '_').replace(/\&/g, '_') ===
       params.category
   ).categoryId
   if (!categoryId) throw new Error('Category id not foud')
