+++ apps/masterbots.ai/components/b/bot-details.tsx
