--- apps/masterbots.ai/components/browse/browse-list.tsx
@@ -1,77 +1,0 @@
-'use client'
-
-import React from 'react'
-import { debounce } from 'lodash'
-import type { Thread } from '@repo/mb-genql'
-import { useBrowse } from '@/hooks/use-browse'
-import { getBrowseThreads } from '@/services/hasura'
-import BrowseListItem from './browse-list-item'
-
-const PAGE_SIZE = 50
-
-export default function BrowseList({ initialThreads }: BrowseListProps) {
-  const { keyword, tab } = useBrowse()
-  const [threads, setThreads] = React.useState<Thread[]>(initialThreads)
-  const [filteredThreads, setFilteredThreads] =
-    React.useState<Thread[]>(initialThreads)
-  const [loading, setLoading] = React.useState<boolean>(false)
-  const [count, setCount] = React.useState<number>(initialThreads.length)
-
-  const verifyKeyword = () => {
-    if (!keyword) {
-      setFilteredThreads(threads)
-    } else {
-      debounce(() => {
-        // TODO: Improve thread messages architecture to implement dynamic search to show only the thread title (first message on thread)
-        // fetchThreads(keyword, tab)
-        setFilteredThreads(
-          threads.filter((thread: Thread) =>
-            thread.messages[0]?.content
-              .toLowerCase()
-              .includes(keyword.toLowerCase())
-          )
-        )
-        // ? Average time of human reaction is 230ms
-      }, 230)()
-    }
-  }
-
-  const loadMore = async () => {
-    console.log('🟡 Loading More Content')
-    setLoading(true)
-
-    const moreThreads = await getBrowseThreads({
-      categoryId: tab,
-      offset: threads.length,
-      limit: PAGE_SIZE
-    })
-
-    setThreads(prevState => [...prevState, ...moreThreads])
-    setCount(moreThreads.length)
-    setLoading(false)
-  }
-
-  React.useEffect(() => {
-    verifyKeyword()
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [keyword, threads])
-
-  return (
-    <div className="w-full py-5 flex flex-col gap-3">
-      {filteredThreads.map((thread: Thread, key) => (
-        <BrowseListItem
-          hasMore={count === PAGE_SIZE}
-          isLast={key === filteredThreads.length - 1}
-          key={key}
-          loadMore={loadMore}
-          loading={loading}
-          thread={thread}
-        />
-      ))}
-    </div>
-  )
-}
-
-type BrowseListProps = {
-  initialThreads: Thread[]
-}
