+++ apps/masterbots.ai/components/shared/thread-dialog/index.ts
@@ -1,0 +1,1 @@
+export * from './thread-dialog'
