+++ apps/masterbots.ai/components/shared/thread-dialog/thread-accordion.tsx
